+++ src/mailman/database/schema/mm_20121015000000.py
@@ -17,10 +17,12 @@
 
 """3.0b2 -> 3.0b3 schema migrations.
 
-* bans.mailing_list -> bans.list_id
+Renamed:
+ * bans.mailing_list -> bans.list_id
 
-Added:
-* mailinglist.style
+Removed:
+ * mailinglist.new_member_options
+ * mailinglist.send_remindersn
 """
 
 from __future__ import absolute_import, print_function, unicode_literals
@@ -91,5 +93,6 @@ def upgrade_postgres(database, store, version, module_path):
             """.format(_make_listid(mailing_list), id))
     store.execute('ALTER TABLE ban DROP COLUMN mailing_list;')
     store.execute('ALTER TABLE mailinglist DROP COLUMN new_member_options;')
+    store.execute('ALTER TABLE mailinglist DROP COLUMN send_reminders;')
     # Record the migration in the version table.
     database.load_schema(store, version, None, module_path)