+++ src/mailman/docs/NEWS.rst
@@ -73,8 +73,8 @@ Database
 --------
  * The `ban` table now uses list-ids to cross-reference the mailing list,
    since these cannot change even if the mailing list is moved or renamed.
- * The column `mailinglist.new_member_options` was unused, and has been
-   removed.
+ * The columns `mailinglist.new_member_options` and
+   `mailinglist.send_reminders` were unused, and have been removed.
 
 Interfaces
 ----------